MST v4.10 Released

+++ Assets/MasterServerToolkit/MasterServer/Scripts/Modules/Authentication/MstAuthServer.cs
@@ -31,7 +31,7 @@ namespace MasterServerToolkit.MasterServer
                 return;
             }
 
-            connection.SendMessage((ushort)MstOpCodes.GetPeerAccountInfo, peerId, (status, response) =>
+            connection.SendMessage(MstOpCodes.GetPeerAccountInfo, peerId, (status, response) =>
             {
                 if (status != ResponseStatus.Success)
                 {
